Prototype for InvokeVirtual optimizations

+++ core/src/main/java/de/mirkosertic/bytecoder/core/optimizer/DeleteCopyToUnusedPHI.java
@@ -15,25 +15,27 @@
  */
 package de.mirkosertic.bytecoder.core.optimizer;
 
+import de.mirkosertic.bytecoder.core.ir.Constant;
 import de.mirkosertic.bytecoder.core.ir.ControlTokenConsumer;
 import de.mirkosertic.bytecoder.core.ir.Copy;
 import de.mirkosertic.bytecoder.core.ir.Graph;
 import de.mirkosertic.bytecoder.core.ir.Node;
+import de.mirkosertic.bytecoder.core.ir.PHI;
 import de.mirkosertic.bytecoder.core.ir.ResolvedMethod;
 import de.mirkosertic.bytecoder.core.ir.Variable;
 import de.mirkosertic.bytecoder.core.parser.CompileUnit;
 
-public class DeleteRedundantVariables implements Optimizer {
+public class DeleteCopyToUnusedPHI implements Optimizer {
 
     private final NodePatternMatcher patternMatcher;
 
-    public DeleteRedundantVariables() {
+    public DeleteCopyToUnusedPHI() {
         patternMatcher = new NodePatternMatcher(
                 NodePredicates.ofType(Copy.class),
                 NodePredicates.singlePredWithForwardEdge(),
                 NodePredicates.singleSuccWithForwardEdge(),
-                NodePredicates.incomingDataFlows(NodePredicates.item(0, NodePredicates.itemOfType(Variable.class))),
-                NodePredicates.outgoingDataFlows(NodePredicates.item(0, NodePredicates.itemOfType(Variable.class)))
+                NodePredicates.incomingDataFlows(NodePredicates.item(0, NodePredicates.itemOfType(Variable.class).or(NodePredicates.itemOfType(Constant.class)))),
+                NodePredicates.outgoingDataFlows(NodePredicates.item(0, NodePredicates.itemOfType(PHI.class)))
         );
     }
 
@@ -48,20 +50,20 @@ public class DeleteRedundantVariables implements Optimizer {
                 final Node incoming = node.incomingDataFlows[0];
                 final Node outgoing = node.outgoingFlows[0];
 
-                g.remapDataFlow(outgoing, incoming);
+                if (outgoing.outgoingFlows.length == 0 && outgoing.incomingDataFlows.length == 1) {
 
-                incoming.removeFromIncomingData(node);
+                    incoming.removeFromIncomingData(node);
+                    outgoing.clearIncomingData();
 
-                outgoing.clearIncomingData();
+                    final ControlTokenConsumer pred = ct.controlComingFrom.iterator().next();
+                    final ControlTokenConsumer succ = ct.controlFlowsTo.values().iterator().next();
 
-                final ControlTokenConsumer pred = ct.controlComingFrom.iterator().next();
-                final ControlTokenConsumer succ = ct.controlFlowsTo.values().iterator().next();
+                    pred.remapControlFlowTo(ct, succ);
 
-                pred.remapControlFlowTo(ct, succ);
+                    g.deleteNode(node);
 
-                g.deleteNode(node);
-
-                return true;
+                    return true;
+                }
             }
         }
